pkg/torcx: move to internal package
