+++ ftests/torcx_image_list_test.go
@@ -25,7 +25,7 @@ import (
 	"testing"
 
 	"github.com/coreos/torcx/internal/cli"
-	"github.com/coreos/torcx/pkg/torcx"
+	"github.com/coreos/torcx/internal/torcx"
 )
 
 func TestImageListEmpty(t *testing.T) {
