+++ internal/cli/image_list.go
@@ -18,7 +18,7 @@ import (
 	"encoding/json"
 	"os"
 
-	"github.com/coreos/torcx/pkg/torcx"
+	"github.com/coreos/torcx/internal/torcx"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
