+++ internal/cli/profile.go
@@ -19,7 +19,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
-	"github.com/coreos/torcx/pkg/torcx"
+	"github.com/coreos/torcx/internal/torcx"
 )
 
 var (
