+++ internal/cli/profile_new.go
@@ -25,7 +25,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
-	"github.com/coreos/torcx/pkg/torcx"
+	"github.com/coreos/torcx/internal/torcx"
 )
 
 var (
