+++ internal/cli/torcx-generator.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/Sirupsen/logrus"
 	logrus_syslog "github.com/Sirupsen/logrus/hooks/syslog"
-	"github.com/coreos/torcx/pkg/torcx"
+	"github.com/coreos/torcx/internal/torcx"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
